Extract game board to seperate component
