+++ src/css/content.css
@@ -42,8 +42,8 @@
 }
 
 .feature-img-container {
-  max-height: 200px;
-  max-width: 200px;
+  width: 200px;
+  height: 200px;
   border-radius: 50%;
   flex-shrink: 0;
   overflow: hidden;
