+++ src/css/images.css
@@ -1,3 +1,3 @@
 .feature-img {
-  max-width: 400px;
+  max-height: 200px;
 }
