+++ src/css/tick-tack-toe.css
@@ -1,9 +1,8 @@
-.tick-tack-toe,
+.tick-tack-toe-page,
 .tick-tack-toe-board {
   display: flex;
   flex-direction: column;
   align-items: center;
-  text-align: justify;
 }
 
 .tile-row {
